Actually implement the code to use the translation stuff.  This means the translation stuff is now functional.  THIS IS NOT DONE THE SAME WAY AS IRCII DOES IT.
ircII implements translation on the output side of the equation, from which
our code has deviated significantly.  Upon reflection, the most common request
for translation was not for OUTPUT translation, but for SERVER translation.
Therefore, I have implemented translation's operations on the server side
of the equation instead of the server side.  If this ends up being the wrong
thing, I reserve the right to go back and "fix" this later.

Anyhow, the translation(s) in force will occur when data is recieved or
transmitted to the server, and not otherwise.  Feedback is appreciated.