This extends /encrypt to add a new argument, which will be treated as a program with which to filter messages through to be encrypted/decrypted.
The protocol so far is:  One line in text mode, with the key as the first
word.  Followed (starting on the second line and ending at EOF), by the
encrypted message in binary.  Binary messages returned are automatically
quoted and unquoted for transmission over the wire.  Ascii Armoring
if performed by the program will "bypass" the quoting features.

This patch also alters the way floating point numbers are returned.
Before, canon_number was passed over the textual value.  Now, ftoa()
reads the correct way to do it from FLOATING_POINT_MATH.