Two new functions in ircsig.c: "block_signal" and "unblock_signal" which use sigprocmask() which I hope everybody has!
Based on a discussion with |Rain|, block the SIGCHLD signal before we
go waitpid()ing, and then unblock it when we're done.  This cleans up a
race condition that might happen if two of our child processes die in the
same time slice; we would only get one call to our SIGCHLD handler, but we
would have two children to reap; "dead_children_processes" would then be
a negative value and bad things would occur.  Now we handle SIGCHLD as a
condition rather than a discrete event; whenever we get a SIGCHLD, the next
time through a sequence point we block SIGCHLD, reap our children, reset
"dead_children_processes" to 0 and then unblock SIGCHLD and go on our way.