Fix a bug found by larne who deserves a lot of credit for pestering me until i finally figured out what caused it.
The bug:  When windows are removed from the screen (via kill or hide), the
space it leaves behind is distributed to the other windows; however, the
window's top-of-window pointer is not moved until the next update_all_windows.
If the window is vacated from the screen before that can occur, the window
can think it is big, but the top of the window will not be pointing at the
place it should be; this can cause output to be accepted onto the window
without it logically scrolling.  When you swap the window back in, it has
output "below" the bottom fo the window.

The evidence:
	/window new hide swap last
	[wait for output to appear in the hidden window]
	/window swap last

You can use /window 1 describe and look at the "Geometry Info" to see the
problem.  All of this could be avoided by using /window new_hide. ;-)